Final Version
Fixed the width of Hero Photo (Needed to offset the columns so it only
spanned 10 instead of 12 cols).

Fixed Navbar by creating another container within the navbar and using
columns to offset it equal to the same offset as the rest of the site).

Fixed some small items (spacing of rows, adding an em dash, etc.)

Cleaned up the HTML with comments and spacing.